Added Technical doc and ppt

+++ backend/src/main/java/com/backend/educonsultancy_backend/controllers/ContactController.java
@@ -66,7 +66,7 @@ public class ContactController {
 
     // Get contacts by user ID
     @GetMapping("/user/{userId}")
-    public ResponseEntity<List<Contact>> getContactsByUserId(@PathVariable("userId") Long userId) {
+    public ResponseEntity<List<Contact>> getContactsByUserId(@PathVariable("userId") Integer userId) {
         List<Contact> contacts = contactService.getContactsByUserId(userId);
         if (contacts.isEmpty()) {
             return new ResponseEntity<>(contacts, HttpStatus.OK);
